Refactor move info page to info folder
